finished tutor side also reoginzed files
