+++ frontend/src/screens/teacherScreensOnly/CreateLesson.tsx
@@ -16,47 +16,48 @@ import { Ionicons } from "@expo/vector-icons";
 import { useNavigation } from "@react-navigation/native";
 import { StackNavigationProp } from "@react-navigation/stack";
 
-import { COLORS } from "../styles";
-import { AuthContext } from "../hooks/AuthContext";
-import { apiService, NewCourse, Category } from "../services/api";
-import { RootStackParamList } from "../types";
+import { COLORS } from "../../styles";
+import { AuthContext } from "../../hooks/AuthContext";
+import { apiService, NewLesson, Course } from "../../services/api";
 
-type CreateCourseNavigationProp = StackNavigationProp<RootStackParamList, "CreateCourse">;
+type CreateLessonNavigationProp = StackNavigationProp<Record<string, object | undefined>>;
 
-export default function CreateCourse() {
+export default function CreateLesson() {
   const { user } = useContext(AuthContext);
-  const navigation = useNavigation<CreateCourseNavigationProp>();
+  const navigation = useNavigation<CreateLessonNavigationProp>();
   const [formData, setFormData] = useState({
     title: "",
     description: "",
-    categoryId: "",
+    courseId: "",
   });
-  const [categories, setCategories] = useState<Category[]>([]);
+  const [courses, setCourses] = useState<Course[]>([]);
   const [loading, setLoading] = useState(false);
-  const [loadingCategories, setLoadingCategories] = useState(true);
-  const [showCategoryModal, setShowCategoryModal] = useState(false);
-  const [selectedCategory, setSelectedCategory] = useState<Category | null>(null);
+  const [loadingCourses, setLoadingCourses] = useState(true);
+  const [showCourseModal, setShowCourseModal] = useState(false);
+  const [selectedCourse, setSelectedCourse] = useState<Course | null>(null);
   const [errors, setErrors] = useState({
     title: "",
     description: "",
-    categoryId: "",
+    courseId: "",
   });
 
   const styles = getStyles();
 
   useEffect(() => {
-    loadCategories();
+    loadCourses();
   }, []);
 
-  const loadCategories = async () => {
+  const loadCourses = async () => {
+    if (!user || !user.isTeacher) return;
+
     try {
-      const categoriesData = await apiService.getAllCategories();
-      setCategories(categoriesData);
+      const coursesData = await apiService.getCoursesByTeacher(user.id);
+      setCourses(coursesData);
     } catch (error) {
-      console.error("Error loading categories:", error);
-      Alert.alert("Error", "Failed to load categories. Please try again.");
+      console.error("Error loading courses:", error);
+      Alert.alert("Error", "Failed to load courses. Please try again.");
     } finally {
-      setLoadingCategories(false);
+      setLoadingCourses(false);
     }
   };
 
@@ -64,7 +65,7 @@ export default function CreateCourse() {
     const newErrors = {
       title: "",
       description: "",
-      categoryId: "",
+      courseId: "",
     };
 
     if (!formData.title.trim()) {
@@ -75,8 +76,8 @@ export default function CreateCourse() {
       newErrors.description = "Description is required";
     }
 
-    if (!formData.categoryId) {
-      newErrors.categoryId = "Category is required";
+    if (!formData.courseId) {
+      newErrors.courseId = "Course is required";
     }
 
     setErrors(newErrors);
@@ -85,7 +86,7 @@ export default function CreateCourse() {
 
   const handleSubmit = async () => {
     if (!user || !user.isTeacher) {
-      Alert.alert("Error", "Only teachers can create courses");
+      Alert.alert("Error", "Only teachers can create lessons");
       return;
     }
 
@@ -96,35 +97,38 @@ export default function CreateCourse() {
     setLoading(true);
 
     try {
-      const courseData: NewCourse = {
+      const lessonData: NewLesson = {
         teacher_id: user.id,
-        category_id: parseInt(formData.categoryId),
+        course_id: parseInt(formData.courseId),
         title: formData.title.trim(),
         description: formData.description.trim(),
+        video_url: "", // Empty for now, will be added via file upload later
       };
 
-      await apiService.createCourse(courseData);
+      await apiService.createLesson(lessonData);
 
-      Alert.alert("Success", "Course created successfully!", [
+      Alert.alert("Success", "Lesson created successfully!", [
         {
           text: "OK",
           onPress: () => {
+            // Reset form
             setFormData({
               title: "",
               description: "",
-              categoryId: "",
+              courseId: "",
             });
+            setSelectedCourse(null);
             setErrors({
               title: "",
               description: "",
-              categoryId: "",
+              courseId: "",
             });
           },
         },
       ]);
     } catch (error) {
-      console.error("Error creating course:", error);
-      Alert.alert("Error", error instanceof Error ? error.message : "Failed to create course. Please try again.");
+      console.error("Error creating lesson:", error);
+      Alert.alert("Error", error instanceof Error ? error.message : "Failed to create lesson. Please try again.");
     } finally {
       setLoading(false);
     }
@@ -150,7 +154,7 @@ export default function CreateCourse() {
         <View style={styles.errorContainer}>
           <Ionicons name="warning-outline" size={64} color={COLORS.error} />
           <Text style={styles.errorTitle}>Access Denied</Text>
-          <Text style={styles.errorText}>Only teachers can create courses.</Text>
+          <Text style={styles.errorText}>Only teachers can create lessons.</Text>
         </View>
       </SafeAreaView>
     );
@@ -163,24 +167,24 @@ export default function CreateCourse() {
           <Ionicons name="arrow-back" size={24} color={COLORS.black} />
         </TouchableOpacity>
         <View style={styles.headerTitleContainer}>
-          <Text style={styles.headerTitleText}>Create Course</Text>
+          <Text style={styles.headerTitleText}>Create Lesson</Text>
         </View>
         <View style={styles.headerSpacer} />
       </View>
 
       <ScrollView style={styles.scrollView} showsVerticalScrollIndicator={false}>
         <View style={styles.header}>
-          <Ionicons name="school-outline" size={32} color={COLORS.purple} />
-          <Text style={styles.headerTitle}>Create New Course</Text>
-          <Text style={styles.headerSubtitle}>Create a course to organize your lessons</Text>
+          <Ionicons name="book-outline" size={32} color={COLORS.purple} />
+          <Text style={styles.headerTitle}>Create New Lesson</Text>
+          <Text style={styles.headerSubtitle}>Create lesson content (video upload coming soon)</Text>
         </View>
 
         <View style={styles.form}>
           <View style={styles.inputGroup}>
-            <Text style={styles.label}>Course Title *</Text>
+            <Text style={styles.label}>Lesson Title *</Text>
             <TextInput
               style={[styles.input, errors.title ? styles.inputError : null]}
-              placeholder="Enter course title..."
+              placeholder="Enter lesson title..."
               value={formData.title}
               onChangeText={(value) => updateFormData("title", value)}
               maxLength={100}
@@ -190,51 +194,53 @@ export default function CreateCourse() {
           </View>
 
           <View style={styles.inputGroup}>
-            <Text style={styles.label}>Category *</Text>
-            {loadingCategories ? (
+            <Text style={styles.label}>Course *</Text>
+            {loadingCourses ? (
               <View style={styles.loadingContainer}>
                 <ActivityIndicator color={COLORS.purple} />
-                <Text style={styles.loadingText}>Loading categories...</Text>
+                <Text style={styles.loadingText}>Loading courses...</Text>
+              </View>
+            ) : courses.length === 0 ? (
+              <View style={styles.noCourseContainer}>
+                <Text style={styles.noCourseText}>No courses found. Please create a course first.</Text>
               </View>
             ) : (
               <TouchableOpacity
-                style={[styles.pickerContainer, errors.categoryId ? styles.inputError : null]}
-                onPress={() => setShowCategoryModal(true)}
+                style={[styles.pickerContainer, errors.courseId ? styles.inputError : null]}
+                onPress={() => setShowCourseModal(true)}
               >
-                <Text style={[styles.pickerText, !selectedCategory && styles.placeholderText]}>
-                  {selectedCategory ? selectedCategory.title : "Select a category..."}
+                <Text style={[styles.pickerText, !selectedCourse && styles.placeholderText]}>
+                  {selectedCourse ? selectedCourse.title : "Select a course..."}
                 </Text>
                 <Ionicons name="chevron-down" size={20} color={COLORS.gray} />
               </TouchableOpacity>
             )}
-            {errors.categoryId ? <Text style={styles.errorText}>{errors.categoryId}</Text> : null}
+            {errors.courseId ? <Text style={styles.errorText}>{errors.courseId}</Text> : null}
           </View>
 
-          <Modal visible={showCategoryModal} transparent animationType="slide">
+          <Modal visible={showCourseModal} transparent animationType="slide">
             <View style={styles.modalOverlay}>
               <View style={styles.modalContent}>
                 <View style={styles.modalHeader}>
-                  <Text style={styles.modalTitle}>Select Category</Text>
-                  <TouchableOpacity onPress={() => setShowCategoryModal(false)}>
+                  <Text style={styles.modalTitle}>Select Course</Text>
+                  <TouchableOpacity onPress={() => setShowCourseModal(false)}>
                     <Ionicons name="close" size={24} color={COLORS.black} />
                   </TouchableOpacity>
                 </View>
                 <FlatList
-                  data={categories}
+                  data={courses}
                   keyExtractor={(item) => item.id.toString()}
                   renderItem={({ item }) => (
                     <TouchableOpacity
-                      style={styles.categoryItem}
+                      style={styles.courseItem}
                       onPress={() => {
-                        setSelectedCategory(item);
-                        updateFormData("categoryId", item.id.toString());
-                        setShowCategoryModal(false);
+                        setSelectedCourse(item);
+                        updateFormData("courseId", item.id.toString());
+                        setShowCourseModal(false);
                       }}
                     >
-                      <Text style={styles.categoryItemText}>{item.title}</Text>
-                      {selectedCategory?.id === item.id && (
-                        <Ionicons name="checkmark" size={20} color={COLORS.purple} />
-                      )}
+                      <Text style={styles.courseItemText}>{item.title}</Text>
+                      {selectedCourse?.id === item.id && <Ionicons name="checkmark" size={20} color={COLORS.purple} />}
                     </TouchableOpacity>
                   )}
                 />
@@ -246,7 +252,7 @@ export default function CreateCourse() {
             <Text style={styles.label}>Description *</Text>
             <TextInput
               style={[styles.textArea, errors.description ? styles.inputError : null]}
-              placeholder="Describe what this course will cover..."
+              placeholder="Describe what students will learn in this lesson..."
               value={formData.description}
               onChangeText={(value) => updateFormData("description", value)}
               multiline
@@ -268,7 +274,7 @@ export default function CreateCourse() {
             ) : (
               <>
                 <Ionicons name="add-circle-outline" size={20} color={COLORS.white} />
-                <Text style={styles.submitButtonText}>Create Course</Text>
+                <Text style={styles.submitButtonText}>Create Lesson</Text>
               </>
             )}
           </TouchableOpacity>
@@ -298,6 +304,7 @@ function getStyles() {
       fontWeight: "bold",
       color: COLORS.black,
       marginTop: 12,
+      textAlign: "center",
     },
     headerSubtitle: {
       fontSize: 16,
@@ -319,35 +326,75 @@ function getStyles() {
     },
     input: {
       borderWidth: 1,
-      borderColor: COLORS.lightGray,
+      borderColor: "#E5E5EA",
       borderRadius: 12,
-      padding: 16,
+      paddingHorizontal: 16,
+      paddingVertical: 12,
       fontSize: 16,
       color: COLORS.black,
       backgroundColor: COLORS.white,
     },
     textArea: {
       borderWidth: 1,
-      borderColor: COLORS.lightGray,
+      borderColor: "#E5E5EA",
       borderRadius: 12,
-      padding: 16,
+      paddingHorizontal: 16,
+      paddingVertical: 12,
       fontSize: 16,
       color: COLORS.black,
       backgroundColor: COLORS.white,
-      minHeight: 120,
+      minHeight: 100,
       textAlignVertical: "top",
     },
-    pickerContainer: {
-      borderWidth: 1,
-      borderColor: COLORS.lightGray,
+    inputError: {
+      borderColor: COLORS.error,
+    },
+    errorText: {
+      color: COLORS.error,
+      fontSize: 14,
+      marginTop: 4,
+    },
+    characterCount: {
+      fontSize: 12,
+      color: COLORS.gray,
+      textAlign: "right",
+      marginTop: 4,
+    },
+    helperText: {
+      fontSize: 12,
+      color: COLORS.gray,
+      marginTop: 4,
+    },
+    submitButton: {
+      backgroundColor: COLORS.purple,
       borderRadius: 12,
-      backgroundColor: COLORS.white,
+      paddingVertical: 16,
       flexDirection: "row",
       alignItems: "center",
-      paddingRight: 16,
+      justifyContent: "center",
+      marginTop: 20,
     },
-    picker: {
-      height: 50,
+    submitButtonDisabled: {
+      backgroundColor: COLORS.gray,
+    },
+    submitButtonText: {
+      color: COLORS.white,
+      fontSize: 16,
+      fontWeight: "600",
+      marginLeft: 8,
+    },
+    errorContainer: {
+      flex: 1,
+      justifyContent: "center",
+      alignItems: "center",
+      padding: 20,
+    },
+    errorTitle: {
+      fontSize: 24,
+      fontWeight: "bold",
+      color: COLORS.black,
+      marginTop: 16,
+      marginBottom: 8,
     },
     loadingContainer: {
       flexDirection: "row",
@@ -361,50 +408,26 @@ function getStyles() {
       fontSize: 16,
       color: COLORS.gray,
     },
-    inputError: {
-      borderColor: COLORS.error,
-    },
-    errorText: {
-      fontSize: 14,
-      color: COLORS.error,
-      marginTop: 4,
-    },
-    characterCount: {
-      fontSize: 12,
-      color: COLORS.gray,
-      textAlign: "right",
-      marginTop: 4,
-    },
-    submitButton: {
-      backgroundColor: COLORS.purple,
+    noCourseContainer: {
       padding: 16,
+      backgroundColor: "#FFF3CD",
       borderRadius: 12,
+      borderWidth: 1,
+      borderColor: "#FFEAA7",
+    },
+    noCourseText: {
+      fontSize: 16,
+      color: "#856404",
+      textAlign: "center",
+    },
+    pickerContainer: {
+      borderWidth: 1,
+      borderColor: "#E5E5EA",
+      borderRadius: 12,
+      backgroundColor: COLORS.white,
       flexDirection: "row",
       alignItems: "center",
-      justifyContent: "center",
-      marginTop: 20,
-    },
-    submitButtonDisabled: {
-      backgroundColor: COLORS.gray,
-    },
-    submitButtonText: {
-      color: COLORS.white,
-      fontSize: 18,
-      fontWeight: "600",
-      marginLeft: 8,
-    },
-    errorContainer: {
-      flex: 1,
-      alignItems: "center",
-      justifyContent: "center",
-      padding: 20,
-    },
-    errorTitle: {
-      fontSize: 24,
-      fontWeight: "bold",
-      color: COLORS.error,
-      marginTop: 16,
-      marginBottom: 8,
+      paddingRight: 16,
     },
     pickerText: {
       fontSize: 16,
@@ -439,7 +462,7 @@ function getStyles() {
       fontWeight: "600",
       color: COLORS.black,
     },
-    categoryItem: {
+    courseItem: {
       flexDirection: "row",
       justifyContent: "space-between",
       alignItems: "center",
@@ -447,7 +470,7 @@ function getStyles() {
       borderBottomWidth: 1,
       borderBottomColor: COLORS.lightGray,
     },
-    categoryItemText: {
+    courseItemText: {
       fontSize: 16,
       color: COLORS.black,
     },
@@ -473,7 +496,7 @@ function getStyles() {
       color: COLORS.black,
     },
     headerSpacer: {
-      width: 40,
+      width: 40, // Same width as back button to center the title
     },
   });
 }
