+++ frontend/src/services/api.ts
@@ -131,6 +131,7 @@ export interface Tutor {
   user_id: number;
   first_name: string;
   last_name: string;
+  username: string;
   email: string;
   phone_number: string;
   category: string;
